+++ src/lib/item/intensify.lua
@@ -11,13 +11,38 @@ local Print
 
 local Random = require 'math_lib'.Random
 local INTENSIFY_COEFFICIENT = {[0] = 1, 2, 3, 4, 6, 8, 10, 14, 18, 22, 31}
+local INTENSIFY_POS = {-2026, -3991}
 
 function Intensify:__call(equipment)
+    -- 檢查有沒有在精煉爐附近
+    local Point = require 'point'
+    local unit_point, pos = Point.GetUnitLoc(equipment.owner_.object_)
+    local is_near = false
+    for i = 1, #INTENSIFY_POS, 2 do
+        pos = Point(INTENSIFY_POS[i], INTENSIFY_POS[i+1])
+        
+        if Point.Distance(unit_point, pos) < 200 then
+            is_near = true
+
+            -- 避免break後的動作不會執行
+            pos:Remove()
+            break
+        end
+
+        pos:Remove()
+    end
+
+    unit_point:Remove()
+
     local Tip = require 'jass_tool'.Tip
 
-    if not NotLimit(equipment) then
-        Tip(equipment.own_player_.object_, {"|cff00ff00提示|r - ", equipment:name(), " 的精鍊值已達上限。"})
+    if not is_near then
+        Tip(equipment.own_player_.object_, "|cff00ff00提示|r - 你不在精煉爐附近。")
+        return false
+    end
 
+    if not NotLimit(equipment) then
+        Tip(equipment.own_player_.object_, {"|cff00ff00提示|r - ", equipment:name(), " 的精煉值已達上限。"})
         return false 
     end
 
@@ -31,7 +56,7 @@ function Intensify:__call(equipment)
 
     -- 精鍊失敗
     if not IsRefine(equipment.intensify_level_) then
-        Tip(equipment.own_player_.object_, {"|cff00ff00提示|r - ", equipment:name(), " 精鍊失敗。"})
+        Tip(equipment.own_player_.object_, {"|cff00ff00提示|r - ", equipment:name(), " 精煉失敗。"})
         
         -- 失敗次數會提高下次強化的花費
         equipment.intensify_fail_times_ = equipment.intensify_fail_times_ + 1
@@ -53,15 +78,19 @@ function Intensify:__call(equipment)
 
     -- 搜尋可被強化的屬性並增加屬性值
     for i = 1, equipment.attribute_count_ do
+        local plus_value = fixed_value + Random(random_max_value)
         if INTENSIFY_ATTRIBUTE[equipment.attribute_[i][1]] then
-            equipment.attribute_[i][2] = equipment.attribute_[i][2] + fixed_value + Random(random_max_value)
+            equipment.attribute_[i][2] = equipment.attribute_[i][2] + plus_value
+            
+            -- 英雄屬性也要更新
+            equipment.owner_:add(equipment.attribute_[i][1], plus_value)
         end
     end
     
     -- 更新屬性
     equipment:Update()
     
-    Tip(equipment.own_player_.object_, {"|cff00ff00提示|r - ", equipment:name(), " 精鍊成功。"})
+    Tip(equipment.own_player_.object_, {"|cff00ff00提示|r - ", equipment:name(), " 精煉成功。"})
 end
 
 NotLimit = function(item)